pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-actor"
version = "0.1.1"
description = "Thin Pipecat wrapper for Deepgram Flux → Gemini/Ollama → Deepgram Aura voice pipeline"
authors = [{name = "Jan Zuiderveld, Grigore Burloiu, Lorena Cocora"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pipecat-ai[deepgram,google,groq,local,silero,local-smart-turn-v3]",
    "pyaec",
    "python-dotenv",
    "pydantic",
    "sounddevice>=0.4",
    "typer>=0.9",
    "watchfiles",
    "transformers[torch]==4.48.2",
    "onnx",
    "onnxruntime",
    "keyboard",
    "python-osc"
]

[project.optional-dependencies]
test = [
    "pytest",
]

[tool.setuptools.packages.find]
where = ["src"]
